start of refactor
