+++ .gitignore
@@ -1,0 +1,12 @@
+node_modules
+dist
+typings
+coverage
+/.vs
+*.user
+npm-debug.log
+*.tgz
+*.log
+/build
+/release
+/mochawesome-reports
