+++ packages/livedoc-mocha/_src/app/BackgroundContext.ts
@@ -1,0 +1,5 @@
+import { ScenarioContext } from "./index";
+
+export class BackgroundContext extends ScenarioContext {
+
+}
