+++ packages/livedoc-mocha/_src/app/FeatureContext.ts
@@ -1,0 +1,6 @@
+export class FeatureContext {
+    filename: string;
+    title: string;
+    description: string;
+    tags: string[];
+}
