+++ packages/livedoc-mocha/_src/app/Row.ts
@@ -1,0 +1,3 @@
+export interface Row {
+    [prop: string]: any;
+}
