+++ packages/livedoc-mocha/_src/app/ScenarioOutlineContext.ts
@@ -1,0 +1,6 @@
+import { ScenarioContext } from "./index";
+import { Row } from "./Row";
+
+export class ScenarioOutlineContext extends ScenarioContext {
+    example: Row;
+}
