+++ packages/livedoc-mocha/_src/test/Background.Spec.ts
@@ -1,5 +1,6 @@
-///<reference path="../app/livedoc.ts" />
+///<reference path="../app/index.ts" />
 import * as Utils from "./Utils";
+import "../app/index";
 
 require('chai').should();
