+++ packages/livedoc-mocha/package.json
@@ -38,8 +38,9 @@
   "devDependencies": {
     "@types/chai": "^3.4.35",
     "@types/mocha": "^2.2.40",
-    "@types/node": "^7.0.5",
+    "@types/node": "^7.0.43",
     "chai": "^3.5.0",
+    "mocha": "^3.5.0",
     "rimraf": "^2.6.1",
     "tslint": "^4.4.2",
     "typescript": "^2.1.6",
